score_cell helper
Score rendering was extracted to the helper method
