+++ app/views/challenges/models/_scores.html.erb
@@ -34,11 +34,7 @@
             <%= source_language %>
           </td>
           <% model_row.target_languages.each do |target_language| %>
-            <%- score = model_row.score(source_language, target_language) -%>
-              <td class="score-cell"
-                style="background-color: <%= score_color(score) %>">
-              <%= number_with_precision(score&.value, precision: 3) || "N/A" %>
-            </td>
+            <%= score_cell model_row.score(source_language, target_language) -%>
           <% end %>
         </tr>
       <% end %>
